+++ Yvand.EntraCP.Tests/GuestAccountsUPNTests.cs
@@ -38,9 +38,9 @@ namespace Yvand.EntraClaimsProvider.Tests
 
         [Test]
         [Repeat(5)]
-        public override void TestAugmentationForUsersMembersOfAllGroups()
+        public override void TestAugmentationOfGoldUsersAgainstRandomGroups()
         {
-            base.TestAugmentationForUsersMembersOfAllGroups();
+            base.TestAugmentationOfGoldUsersAgainstRandomGroups();
         }
     }
 }